fixed binding on users, mentee, log and added role in user response

+++ feature/log/delivery/request.go
@@ -4,14 +4,14 @@ import "github.com/GP2-Group5/Backend/feature/log"
 
 type LogReq struct {
 	Feedback  string `json:"feedback"`
-	Status_id int    `json:"status_id"`
-	Mentee_id int    `json:"mentee_id"`
+	Status_id uint   `json:"status_id"`
+	Mentee_id uint   `json:"mentee_id"`
 }
 
 func toCore(data LogReq) log.LogCore {
 	return log.LogCore{
 		Feedback: data.Feedback,
-		StatusID: uint(data.Status_id),
-		MenteeID: uint(data.Mentee_id),
+		StatusID: data.Status_id,
+		MenteeID: data.Mentee_id,
 	}
 }
