+++ feature/users/delivery/request.go
@@ -8,7 +8,7 @@ type UserReq struct {
 	Full_Name string `json:"full_name" form:"full_name"`
 	Password  string `json:"password" form:"password"`
 	Email     string `json:"email" form:"email"`
-	Team_id   int    `json:"team_id" form:"team_id"`
+	Team_id   uint   `json:"team_id" form:"team_id"`
 	Status    string `json:"status" form:"status"`
 	Role      string `json:"role" form:"role"`
 }
@@ -19,6 +19,6 @@ func toCore(data UserReq) users.UserCore {
 		Email:     data.Email,
 		Password:  data.Password,
 		Role:      data.Role,
-		TeamID:    uint(data.Team_id),
+		TeamID:    data.Team_id,
 	}
 }
